refactor: refactor `get_query_result` (#36057)
